finna/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finna-backend"
version = "0.1.0"
description = "Finna backend foundation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13,<2.0",
  "asyncpg>=0.29,<1.0",
  "email-validator>=2.1,<3.0",
  "fastapi>=0.110,<1.0",
  "httpx>=0.27,<1.0",
  "openpyxl>=3.1,<4.0",
  "pdfplumber>=0.11,<1.0",
  "pytesseract>=0.3.10,<1.0",
  "psycopg[binary]>=3.1,<4.0",
  "pydantic-settings>=2.2,<3.0",
  "pyjwt[crypto]>=2.8,<3.0",
  "python-dateutil>=2.9,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "rapidfuzz>=3.9,<4.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.29,<1.0",
]

[project.optional-dependencies]
dev = [
  "asgi-lifespan>=2.1,<3.0",
  "pytest>=8.2,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "respx>=0.21,<1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
